apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
scapy==2.5.0
requests==2.32.3
cryptography==43.0.0
netdisco==3.0.0
python-nmap==0.7.1
